Added print out of snapshot.value
