+++ finalProjectMA.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:ShowEventsListTableViewController.swift">
+      location = "group:finalProjectMA/ShowEventsTableViewController.swift">
    </FileRef>
    <FileRef
       location = "group:keys.plist">
