+++ finalProjectMA/Event.swift
@@ -7,16 +7,27 @@
 //
 
 import Foundation
+import Firebase
 
 struct Event {
     let name: String
     let time: String
     let location: String
+    let ref: FIRDatabaseReference?
     
     init(name:String, time:String, location:String){
         self.name = name
         self.time = time
         self.location = location
+        self.ref = nil
+    }
+    
+    init(snapshot: FIRDataSnapshot) {
+        let snapshotValue = snapshot.value as! [String: AnyObject]
+        name = snapshotValue["name"] as! String
+        time = snapshotValue["time"] as! String
+        location = snapshotValue["location"] as! String
+        ref = snapshot.ref
     }
     
     func toAnyObject() -> Any {
@@ -26,5 +37,4 @@ struct Event {
             "location": location
         ]
     }
-    
 }
